+++ src/components/header.js
@@ -1,10 +1,7 @@
 import React from "react";
 import Link from "gatsby-link";
 import AUheader, { AUheaderBrand } from "@gov.au/header";
-import AUaccordion from "@gov.au/accordion";
-
 import crest from "../assets/crest.svg";
-import star from "../assets/star_white.svg";
 
 import "./header.scss";
 
@@ -34,30 +31,16 @@ const Header = ({ siteTitle }) => (
 );
 
 const OfficalBanner = () => (
-	<div className="offical-banner">
-		{" "}
-		<AUaccordion
-			header={
-				<div>
-					<span className="offical-logo">
-						<img src={star} alt="Australia star logo" />
-					</span>{" "}
-					<span className="offical-title">Australian Government</span>{" "}
-					<span className="offical-subline">official website</span>
-				</div>
-			}
+	<div className="official-banner">
+		This website is in alpha. We’re testing new designs and a small amount
+		of content. Let us know what you think. Go back to{" "}
+		<a
+			href="https://www.australia.gov.au/"
+			alt="australia.gov.au"
+			target="_blank"
 		>
-			<div className="offical-content">test</div>
-		</AUaccordion>
-		<span className="offical-link">
-			<a
-				href="https://www.dta.gov.au/"
-				alt="DTA website"
-				className="banner-link"
-			>
-				dta.gov.au
-			</a>
-		</span>
+			australia.gov.au
+		</a>
 	</div>
 );